+++ packages/falcon-dev-tools/eslint-config-falcon/index.js
@@ -4,10 +4,7 @@ module.exports = {
   plugins: ['react', 'import', 'prettier', 'jsdoc'],
   settings: {
     'import/parser': 'babel-eslint',
-    'import/resolver': {
-      node: true,
-      'eslint-import-resolver-typescript': true
-    },
+    'import/resolver': { node: true },
     jsdoc: {
       tagNamePreference: {
         return: 'returns'
@@ -23,12 +20,7 @@ module.exports = {
     'class-methods-use-this': 'off',
     'consistent-return': 'off',
     'global-require': 'off',
-    'id-length': [
-      1,
-      {
-        exceptions: ['_', 'e', 't', 'x', 'p', 'm', 'i', 'j']
-      }
-    ],
+    'id-length': [1, { exceptions: ['_', 'e', 't', 'x', 'p', 'm', 'i', 'j'] }],
     'import/default': 'off',
     'import/no-duplicates': 'off',
     'import/named': 'off',
@@ -38,20 +30,13 @@ module.exports = {
     'import/extensions': ['off', 'never'],
     'import/no-extraneous-dependencies': ['error', { devDependencies: true }],
     'import/prefer-default-export': 'off',
-    'jsx-a11y/anchor-is-valid': [
-      'off',
-      {
-        components: ['Link']
-      }
-    ],
+    'jsx-a11y/anchor-is-valid': ['off', { components: ['Link'] }],
     'jsx-a11y/click-events-have-key-events': 'off',
     'jsx-a11y/label-has-for': [
       2,
       {
         components: ['Label'],
-        required: {
-          every: ['id']
-        },
+        required: { every: ['id'] },
         allowChildren: false
       }
     ],
@@ -68,9 +53,7 @@ module.exports = {
         array: false,
         object: true
       },
-      {
-        enforceForRenamedProperties: false
-      }
+      { enforceForRenamedProperties: false }
     ],
     'prettier/prettier': [
       'error',
@@ -80,21 +63,14 @@ module.exports = {
         printWidth: 120
       }
     ],
-    'react/jsx-filename-extension': [
-      1,
-      {
-        extensions: ['.js', '.jsx', '.tsx']
-      }
-    ],
+    'react/jsx-filename-extension': [1, { extensions: ['.js', '.jsx', '.tsx'] }],
+    'react/jsx-one-expression-per-line': 'off',
     'react/no-danger': 'off',
     'react/no-multi-comp': 'off',
-    // problem with redux-connect decorator so we cannot use required on props,
-    // therefore everything would need to be defined in default props
     'react/require-default-props': ['off'],
     'react/no-access-state-in-setstate': 'error',
     'react/prop-types': 'off',
     'react/destructuring-assignment': 'off',
-
     'jsdoc/check-alignment': 1,
     'jsdoc/check-indentation': 1,
     'jsdoc/check-param-names': 1,
@@ -122,8 +98,8 @@ module.exports = {
   overrides: [
     {
       files: ['**/*.ts', '**/*.tsx'],
-      parser: 'typescript-eslint-parser',
-      plugins: ['typescript'],
+      parser: '@typescript-eslint/parser',
+      plugins: ['@typescript-eslint'],
       rules: {
         'no-undef': 'off',
         'no-unused-vars': 'off',