+++ packages/falcon-server-env/src/index.ts
@@ -1,9 +1,10 @@
 export { InMemoryLRUCache } from 'apollo-server-caching';
-export { default as ApiDataSource } from './models/ApiDataSource';
-export { default as Extension } from './models/Extension';
-export { default as EndpointManager } from './models/EndpointManager';
+export * from './models/ApiDataSource';
+export * from './models/Extension';
+export * from './models/EndpointManager';
 export { default as ContextHTTPCache } from './cache/ContextHTTPCache';
 export { default as Cache } from './cache/Cache';
 export * from './types';
 export { default as htmlHelpers } from './helpers/htmlHelpers';
 export { Events } from './events';
+export * from './authorization';